# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MetalMark API
  description: Self-hosted personal finance (Monarch-style). LAN/VPN only.
  version: 0.1.0
paths:
...
  /owners:
    get:
      tags:
      - owners
      summary: List Owners
      operationId: list_owners_owners_get
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OwnerOut'
                title: Response List Owners Owners Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - owners
      summary: Create Owner
      operationId: create_owner_owners_post
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /owners/{owner_id}:
    patch:
      tags:
      - owners
      summary: Update Owner
      operationId: update_owner_owners__owner_id__patch
      parameters:
      - name: owner_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Owner Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - owners
      summary: Delete Owner
      description: 'Delete an owner, moving its rows to ``reassign_to`` (default: Shared).


        Owner-only: the operation silently re-attributes existing ledger history, which

        is a household-shape decision, not an everyday edit.'
      operationId: delete_owner_owners__owner_id__delete
      parameters:
      - name: owner_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Owner Id
      - name: reassign_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Reassign To
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerDeleteResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /reports/net-worth:
    get:
      tags:
      - reports
      summary: Net Worth Series
      operationId: net_worth_series_reports_net_worth_get
      parameters:
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Start
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date
          title: End
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetWorthSeries'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/cash-flow:
    get:
      tags:
      - reports
      summary: Cash Flow
      operationId: cash_flow_reports_cash_flow_get
      parameters:
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Start
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date
          title: End
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CashFlowPoint'
                title: Response Cash Flow Reports Cash Flow Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/spending:
    get:
      tags:
      - reports
      summary: Spending
      operationId: spending_reports_spending_get
      parameters:
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Start
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date
          title: End
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpendingReport'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CashFlowPoint:
      properties:
        month:
          type: string
          title: Month
        income:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Income
        expense:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Expense
        net:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Net
      type: object
      required:
      - month
      - income
      - expense
      - net
      title: CashFlowPoint
...
    CategorySpendRow:
      properties:
        category_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Category Id
        category_name:
          type: string
          title: Category Name
        total:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total
      type: object
      required:
      - category_id
      - category_name
      - total
      title: CategorySpendRow
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NetWorthPoint:
      properties:
        date:
          type: string
          format: date
          title: Date
        net_worth:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Net Worth
      type: object
      required:
      - date
      - net_worth
      title: NetWorthPoint
    NetWorthSeries:
      properties:
        base_currency:
          type: string
          title: Base Currency
        points:
          items:
            $ref: '#/components/schemas/NetWorthPoint'
          type: array
          title: Points
        delta_net_worth:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Delta Net Worth
        net_cash_flow:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Net Cash Flow
        currency_revaluation:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Currency Revaluation
        attribution:
          type: string
          title: Attribution
          default: account
      type: object
      required:
      - base_currency
      - points
      - delta_net_worth
      - net_cash_flow
      - currency_revaluation
      title: NetWorthSeries
    OwnerCreate:
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 1
          title: Name
        sort:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sort
      type: object
      required:
      - name
      title: OwnerCreate
    OwnerDeleteResult:
      properties:
        reassigned_accounts:
          type: integer
          title: Reassigned Accounts
        reassigned_transactions:
          type: integer
          title: Reassigned Transactions
        reassigned_splits:
          type: integer
          title: Reassigned Splits
      type: object
      required:
      - reassigned_accounts
      - reassigned_transactions
      - reassigned_splits
      title: OwnerDeleteResult
      description: 'What the delete actually moved — the UI tells the user, rather than

        implying that deleting an owner was free.'
    OwnerOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        kind:
          type: string
          title: Kind
        sort:
          type: integer
          title: Sort
      type: object
      required:
      - id
      - name
      - kind
      - sort
      title: OwnerOut
    OwnerUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 80
            minLength: 1
          - type: 'null'
          title: Name
        sort:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sort
      type: object
      title: OwnerUpdate
...
    SpendingReport:
      properties:
        base_currency:
          type: string
          title: Base Currency
        rows:
          items:
            $ref: '#/components/schemas/CategorySpendRow'
          type: array
          title: Rows
        total:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total
      type: object
      required:
      - base_currency
      - rows
      - total
      title: SpendingReport
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError